get-diet-data/diet-data/requirements.txt
azure-core==1.36.0
azure-functions==1.24.0
azure-storage-blob==12.27.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
isodate==0.7.2
MarkupSafe==3.0.3
pycparser==2.23
python-dotenv==1.2.1
requests==2.32.5
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
